feat(llm): log dynamic content card impression when 50% is shown
